Refactor squashfs handlers to the new class-based implementation
Now we can simplify and put them to the same module, because they can have
a base class and separate class instances in the same module.
