+++ unblob/handlers/__init__.py
@@ -1,16 +1,19 @@
-from typing import Dict, List
+from typing import Dict, List, Type
 
 from ..models import Handler
 from .archive import ar, cab, cpio, tar
-from .filesystem.squashfs import squashfs_v3, squashfs_v4
+from .filesystem import squashfs
 
 
-def _make_handler_map(*handlers: Handler) -> Dict[str, Handler]:
-    return {h.NAME: h for h in handlers}
+def _make_handler_map(*handlers: Type[Handler]) -> Dict[str, Handler]:
+    return {h.NAME: h() for h in handlers}
 
 
 _ALL_MODULES_BY_PRIORITY: List[Dict[str, Handler]] = [
-    _make_handler_map(squashfs_v3, squashfs_v4, cramfs),
+    _make_handler_map(
+        squashfs.SquashFSv3Handler,
+        squashfs.SquashFSv4Handler,
+    ),
     _make_handler_map(
         ar,
         cab,
