--- unblob/handlers/filesystem/squashfs/squashfs_v3.py
@@ -1,93 +1,0 @@
-import io
-import struct
-from typing import List, Union
-
-from dissect.cstruct import cstruct
-from structlog import get_logger
-
-from ....file_utils import round_up
-from ....models import UnknownChunk, ValidChunk
-
-logger = get_logger()
-
-NAME = "squashfs_v3"
-
-YARA_RULE = r"""
-    strings:
-        /**
-        00000000  73 71 73 68 00 00 00 03  00 00 00 00 00 00 00 00  |sqsh............|
-        00000010  00 00 00 00 00 00 00 00  00 00 00 00 00 03 00 00  |................|
-        */
-        $squashfs_v3_magic_be = { 73 71 73 68 [24] 00 03}
-        /**
-        00000000  68 73 71 73 03 00 00 00  00 00 00 00 00 00 00 00  |hsqs............|
-        00000010  00 00 00 00 00 00 00 00  00 00 00 00 03 00 00 00  |................|
-        */
-        $squashfs_v3_magic_le = { 68 73 71 73 [24] 03 00}
-    condition:
-        $squashfs_v3_magic_le or $squashfs_v3_magic_be
-"""
-YARA_MATCH_OFFSET = 0
-
-SUASHFS_V4_HEADER = """
-struct SQUASHFS3_SUPER_BLOCK
-{
-    char   s_magic[4];
-    uint32 inodes;
-    uint32 bytes_used_2;
-    uint32 uid_start_2;
-    uint32 guid_start_2;
-    uint32 inode_table_start_2;
-    uint32 directory_table_start_2;
-    uint16 s_major;
-    uint16 s_minor;
-    uint16 block_size_1;
-    uint16 block_log;
-    uint8  flags;
-    uint8  no_uids;
-    uint8  no_guids;
-    uint32 mkfs_time /* time of filesystem creation */;
-    uint64 root_inode;
-    uint32 block_size;
-    uint32 fragments;
-    uint32 fragment_table_start_2;
-    int64  bytes_used;
-    int64  uid_start;
-    int64  guid_start;
-    int64  inode_table_start;
-    int64  directory_table_start;
-    int64  fragment_table_start;
-    int64  lookup_table_start;
-};
-"""
-cparser_le = cstruct()
-cparser_le.load(SUASHFS_V4_HEADER)
-
-cparser_be = cstruct(endian=">")
-cparser_be.load(SUASHFS_V4_HEADER)
-
-PAD_SIZE = 4096
-BIG_ENDIAN_MAGIC = 0x73717368
-
-
-def calculate_chunk(
-    file: io.BufferedIOBase, start_offset: int
-) -> Union[ValidChunk, UnknownChunk]:
-
-    # read the magic and derive endianness from it
-    magic_bytes = file.read(4)
-    magic = struct.unpack(">I", magic_bytes)[0]
-    cparser = cparser_be if magic == BIG_ENDIAN_MAGIC else cparser_le
-
-    file.seek(start_offset)
-    header = cparser.SQUASHFS3_SUPER_BLOCK(file)
-    logger.debug("Header parsed", header=header)
-
-    size = round_up(header.bytes_used, PAD_SIZE)
-    end_offset = start_offset + size
-
-    return ValidChunk(start_offset=start_offset, end_offset=end_offset)
-
-
-def make_extract_command(inpath: str, outdir: str) -> List[str]:
-    return ["unsquashfs", "-f", "-d", outdir, inpath]
